使用能确定长度的标准int类型

+++ User/bsp/a/bsp_buzzer.c
@@ -10,9 +10,9 @@
 /* Private variables -------------------------------------------------------- */
 /* Private function  -------------------------------------------------------- */
 /* Exported functions ------------------------------------------------------- */
-int BSP_Buzzer_Start(void) { return HAL_TIM_PWM_Start(&htim12, TIM_CHANNEL_4); }
+uint8_t BSP_Buzzer_Start(void) { return HAL_TIM_PWM_Start(&htim12, TIM_CHANNEL_4); }
 
-int BSP_Buzzer_Set(float freq, float duty_cycle) {
+uint8_t BSP_Buzzer_Set(float freq, float duty_cycle) {
   uint16_t pulse = freq * PWM_RESOLUTION;
   __HAL_TIM_SET_COMPARE(&htim12, TIM_CHANNEL_4, pulse);
 
@@ -20,4 +20,4 @@ int BSP_Buzzer_Set(float freq, float duty_cycle) {
   __HAL_TIM_PRESCALER(&htim12, pulse);
   return 0;
 }
-int BSP_Buzzer_Stop(void) { return HAL_TIM_PWM_Stop(&htim12, TIM_CHANNEL_4); }
+uint8_t BSP_Buzzer_Stop(void) { return HAL_TIM_PWM_Stop(&htim12, TIM_CHANNEL_4); }
