+++ User/bsp/a/bsp_fric.c
@@ -9,19 +9,19 @@
 /* Private variables -------------------------------------------------------- */
 /* Private function  -------------------------------------------------------- */
 /* Exported functions ------------------------------------------------------- */
-int BSP_Fric_Start(void) {
+uint8_t BSP_Fric_Start(void) {
   HAL_TIM_PWM_Start(&htim1, TIM_CHANNEL_1);
   BSP_Delay(500);
   HAL_TIM_PWM_Start(&htim1, TIM_CHANNEL_1);
   return 0;
 }
-int BSP_Fric_Set(float duty_cycle) {
+uint8_t BSP_Fric_Set(float duty_cycle) {
   uint16_t pulse = duty_cycle * PWM_RESOLUTION;
   __HAL_TIM_SET_COMPARE(&htim1, TIM_CHANNEL_1, pulse);
   __HAL_TIM_SET_COMPARE(&htim1, TIM_CHANNEL_4, pulse);
 }
 
-int BSP_Fric_Stop(void) {
+uint8_t BSP_Fric_Stop(void) {
   HAL_TIM_PWM_Stop(&htim1, TIM_CHANNEL_1);
   HAL_TIM_PWM_Stop(&htim1, TIM_CHANNEL_4);
 }
