+++ User/bsp/a/bsp_laser.c
@@ -9,12 +9,12 @@
 /* Private variables -------------------------------------------------------- */
 /* Private function  -------------------------------------------------------- */
 /* Exported functions ------------------------------------------------------- */
-int Laser_On(void) {
+uint8_t Laser_On(void) {
   HAL_GPIO_WritePin(LASER_GPIO_Port, LASER_Pin, GPIO_PIN_SET);
   return 0;
 }
 
-int Laser_Off(void) {
+uint8_t Laser_Off(void) {
   HAL_GPIO_WritePin(LASER_GPIO_Port, LASER_Pin, GPIO_PIN_RESET);
   return 0;
 }
