+++ User/bsp/power.h
@@ -6,6 +6,7 @@ extern "C" {
 
 /* Includes ----------------------------------------------------------------- */
 #include <stdbool.h>
+#include <stdint.h>
 
 /* Exported constants ------------------------------------------------------- */
 /* Exported macro ----------------------------------------------------------- */
@@ -18,7 +19,7 @@ typedef enum {
 } BSP_Power_Port_t; /* 电源输出接口 */
 
 /* Exported functions prototypes -------------------------------------------- */
-int BSP_Power_Set(BSP_Power_Port_t port, bool s);
+int8_t BSP_Power_Set(BSP_Power_Port_t port, bool s);
 
 #ifdef __cplusplus
 }
